backend/pyproject.toml
[project]
name = "backend-v2"
version = "0.1.0"
description = "Monitors Media"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "clickhouse-connect>=0.8.15",
    "fastapi>=0.115.11",
    "httpx>=0.28.1",
    "pydantic-settings>=2.8.1",
    "python-dateutil>=2.9.0.post0",
    "uvicorn>=0.34.0",
]
